pyproject.toml
[tool.poetry]
name = "intis"
version = "0.1.0"
description = ""
authors = ["jalelegenda <jakov.milic@yahoo.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["security"], version = "^0.114.1"}
SQLAlchemy = "^2.0.34"
pydantic = "^2.9.1"
PyJWT = "^2.9.0"
Authlib = "^1.3.2"
uvicorn = "^0.30.6"
icalendar = "^5.0.13"
ics = "^0.7.2"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
sqlmodel = "^0.0.22"
python-multipart = "^0.0.9"
httpx = "^0.27.2"
asyncpg = "^0.29.0"
strenum = "^0.4.15"
sortedcontainers = "^2.4.0"
jinja2 = "^3.1.4"
pydantic-settings = "^2.5.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
mypy = "^1.11.2"
ruff = "^0.6.4"
alembic = "^1.13.2"
pytest-asyncio = "^0.24.0"
pytest-mock = "^3.14.0"
polyfactory = "^2.16.2"
cuid2 = "^2.0.1"
pre-commit = "^3.8.0"
coverage = "^7.6.1"
freezegun = "^1.5.1"
pytest-xdist = "^3.6.1"

[tool.pytest.ini_options]
pythonpath = "."
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
testpaths = "tests"
filterwarnings = "ignore::DeprecationWarning:passlib.*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[mypy]
plugins = "pydantic.mypy"
